requirementes.txt
fdk
oci
mysql-connector-python